sparkplug: abstractHandler for Gateway and Sparkplug
